Split RankSelect into several modules
