+++ src/Data/RankSelect/Map.hs
@@ -1,0 +1,3 @@
+module Data.RankSelect.Map where
+
+
